refactored app.rs split the code into multiple files
